Reorganize User module files
