+++ src/server/src/users/controllers/users.controller.ts
@@ -1,7 +1,7 @@
 import { Controller } from "@nestjs/common";
 import { Crud, CrudController } from "@nestjsx/crud";
-import { User } from "../../entities/user.entity";
-import { UsersService } from "../../services/users/users.service";
+import { User } from "../entities/user.entity";
+import { UsersService } from "../services/users.service";
 
 @Crud({
   model: {
