+++ src/server/src/users/services/users.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { TypeOrmCrudService } from "@nestjsx/crud-typeorm";
 import { Repository } from "typeorm";
-import { User } from "../../entities/user.entity";
+import { User } from "../entities/user.entity";
 
 @Injectable()
 export class UsersService extends TypeOrmCrudService<User> {
