+++ src/server/src/users/users.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { UsersController } from "./controllers/users/users.controller";
+import { UsersController } from "./controllers/users.controller";
 import { User } from "./entities/user.entity";
-import { UsersService } from "./services/users/users.service";
+import { UsersService } from "./services/users.service";
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
   controllers: [UsersController],
